[#5] refactor: 타입 정의를 type 파일로 이동
