+++ src/lib/types.ts
@@ -51,12 +51,4 @@ export type PostNew = {
   purchaseLink: string;
   productImgUrl: string;
   content: string;
-}
-
-// 카카오 지도 place 타입
-export interface PlaceType {
-  id: string;
-  position: kakao.maps.LatLng;
-  title: string;
-  address: string;
-}
+};
